Applied additional patches by Daniel Richard G. (20061116)
git-svn-id: svn://svn.code.sf.net/p/yodl/svn/branches/rm_unsigned@92 8f49d26b-2354-4b37-89f9-c6476e23a9d3

+++ yodl/src/hashmap/hashmapkeys.c
@@ -4,7 +4,7 @@ void hashmap_keys(register HashMap *map)
 {
     size_t idx;
 
-    fprintf(stderr, "There are %u active keys\n", map->d_n);
+    fprintf(stderr, "There are %u active keys\n", (unsigned int)map->d_n);
 
     for (idx = 0; idx < map->d_size; idx++)
     {
@@ -16,11 +16,11 @@ void hashmap_keys(register HashMap *map)
             continue;
 
             case REMOVED:
-                fprintf(stderr, "   %3d: <removed>\n", idx);
+                fprintf(stderr, "   %3d: <removed>\n", (int)idx);
             break;
 
             default:
-                fprintf(stderr, "   %3d: `%s'\n", idx, atIdx->d_key);
+                fprintf(stderr, "   %3d: `%s'\n", (int)idx, atIdx->d_key);
             break;
         }
     }
