+++ yodl/src/message/message.c
@@ -14,7 +14,7 @@ void message(char const *fmt,...)
     if (m_message.d_errors >= m_message.d_max_errors)
     {
         fprintf(stderr,
-                "Terminating after %u errors\n", m_message.d_max_errors);
+                "Terminating after %u errors\n", (unsigned int)m_message.d_max_errors);
         exit(1);
     }
 }
