+++ yodl/src/yodlpost/handleref.c
@@ -22,7 +22,7 @@ void handle_ref(long offset, HashItem *item)
     {
         fprintf(global.d_out,
             "<a href=\"%s%02d.%s#%s\">%s</a>",
-            global.d_noext, lp->d_filenr,
+            global.d_noext, (int)lp->d_filenr,
             global.d_ext,
             label,
             lines_at(&global.d_section, lp->d_section));
