+++ yodl/src/yodlpost/handlexmlref.c
@@ -29,7 +29,7 @@ void handle_xml_ref(long offset, HashItem *item)
         fprintf(global.d_out,
             "<a href=\"%s%s%02d#%s\">%s</a>",
             documentbase,
-            global.d_noext, lp->d_filenr,
+            global.d_noext, (int)lp->d_filenr,
             xml_label,
             lines_at(&global.d_section, lp->d_section));
     }
