+++ yodl/src/yodlpost/handlexmlsolvelink.c
@@ -26,7 +26,7 @@ void handle_xml_solvelink(long offset, HashItem *item)
     {
         fprintf(global.d_out, "%s%02d.%s#%s",
                     global.d_noext,
-                    lp->d_filenr,
+                    (int)lp->d_filenr,
                     global.d_ext,
                     xml_label);
     }
